Enhances the binary decode functions to return an enum for the encoding format.
Added a top level binary decoder function to decode "with format", so that
enums for both the opcode as well as the instruction format used by that
opcode are returned as a result. This can be useful where one wants to treat
groups of instruction opcodes similarly based on the format in which they are
encoded.

PiperOrigin-RevId: 617988888
Change-Id: Ic89950da4e4f0b9452268e6a9a271752848d7390