Adding a method to notify core that it should check for interrupts.
This provides a standard interface to notify a core that it should check
to see if there are interrupts to be taken. The default implementation is
empty. It is intended that any code that modifies interrupt flags or
interrupt enable bits for a core should call this function for the core
to determine if an interrupt is available and what to do next.

PiperOrigin-RevId: 633565273
Change-Id: I3349aa396aac8ee475d176d79a6736e02bf0de6d