Added new source/destination, and resource specification in the .isa files.
This specification allows you to specify an operand (resource) as [name].
This translates into an enum entry of k_Name_. This operand/resource
specification is used in cases where it may genereate multiple operands
or resourse references.

For instance, in RiscV push/pop instructions a field 'rlist' is used to
specify a value in the range [4..15] that represents different length
lists of registers. Instead of having to specify the instruction explicitly
for each possible value of 'rlist', the [rlist] specification will allow
new GetResource/GetSource/GetDestination calls in the encoding interface
to return multiple source/destination/resource operands during the decoding
of the instruction.

PiperOrigin-RevId: 692247798
Change-Id: Ie6a08d2df349046a1ccd98c4acf8025a968451d0